Major test refactor. - Reorganize tests into directories, including fixtures - Allow level fixture to be overriden by a pytest class, in addition to module - Add test_resource (including TestResource) to begin Test class hierarchy for resources - Add logged out and test account logged in docker clusters, rearrange sharing and docker fixtures a bit
test_obj_store.py and test_module.py both pass, test_resource.py runs properly (but various failures) - basics seem to work.
