+++ tests/test_resources/test_modules/test_folders/test_folder.py
@@ -22,7 +22,7 @@ def fs_str_rh_fn(folder):
 
 @pytest.mark.clustertest
 def test_from_cluster(cluster):
-    rh.folder(path="./").to(cluster, path="my_new_tests_folder")
+    rh.folder(path="../../../").to(cluster, path="my_new_tests_folder")
     tests_folder = rh.folder(system=cluster, path="my_new_tests_folder")
     assert "my_new_tests_folder/requirements.txt" in tests_folder.ls()
