+++ tests/test_resources/test_resource_sharing.py
@@ -6,9 +6,7 @@ import pytest
 import requests
 
 import runhouse as rh
-from runhouse.globals import configs, rns_client
-
-from tests.conftest import shared_resources, test_account
+from runhouse.globals import rns_client
 
 
 def call_func_with_curl(ip_address, func_name, token, *args):
@@ -52,9 +50,8 @@ def call_cluster_methods(cluster, test_env, valid_token):
 
 
 @pytest.mark.clustertest
-def test_cluster_sharing(shared_resources):
-    current_token = configs.get("token")
-    shared_cluster, shared_function = shared_resources
+def test_cluster_sharing(shared_cluster, shared_function):
+    current_token = rh.configs.get("token")
     # Run commands on cluster with current token
     return_codes = shared_cluster.run_python(
         ["import numpy", "print(numpy.__version__)"]
@@ -73,10 +70,9 @@ def test_cluster_sharing(shared_resources):
 
 
 @pytest.mark.clustertest
-def test_use_shared_cluster_apis(shared_resources, test_env):
+def test_use_shared_cluster_apis(shared_cluster, shared_function, test_env):
     # Should be able to use the shared cluster APIs if given access
-    current_token = configs.get("token")
-    shared_cluster, shared_function = shared_resources
+    current_token = rh.configs.get("token")
 
     # Confirm we can perform cluster actions with the current token
     call_cluster_methods(shared_cluster, test_env, valid_token=True)
@@ -88,17 +84,16 @@ def test_use_shared_cluster_apis(shared_resources, test_env):
         assert "No read or write access to requested resource" in str(e)
 
     # Confirm we cannot perform actions on the cluster with an invalid token
-    configs.set("token", "abc123")
+    rh.configs.set("token", "abc123")
     call_cluster_methods(shared_cluster, test_env, valid_token=False)
 
     # Reset back to valid token
-    configs.set("token", current_token)
+    rh.configs.set("token", current_token)
 
 
 @pytest.mark.clustertest
-def test_use_shared_function_apis(shared_resources):
-    current_token = configs.get("token")
-    shared_cluster, shared_function = shared_resources
+def test_use_shared_function_apis(shared_cluster, shared_function):
+    current_token = rh.configs.get("token")
 
     # Call the function with current valid token
     assert shared_function(2, 2) == 4
@@ -108,14 +103,14 @@ def test_use_shared_function_apis(shared_resources):
     assert reloaded_func(1, 2) == 3
 
     # Use invalid token to confirm no function access
-    configs.set("token", "abc123")
+    rh.configs.set("token", "abc123")
     try:
         shared_function(2, 2) == 4
     except Exception as e:
         assert "Error calling call on server" in str(e)
 
     # Reset back to valid token and confirm we can call function again
-    configs.set("token", current_token)
+    rh.configs.set("token", current_token)
     res = call_func_with_curl(
         shared_cluster.address, shared_function.name, current_token, 1, 2
     )
@@ -123,14 +118,15 @@ def test_use_shared_function_apis(shared_resources):
 
 
 @pytest.mark.clustertest
-def test_running_func_with_cluster_read_access():
+def test_running_func_with_cluster_read_access(
+    test_account, shared_cluster, shared_function
+):
     """Check that a user with read only access to the cluster cannot call a function on that cluster if they do not
     explicitly have access to the function."""
-    current_username = configs.get("username")
-    current_token = configs.get("token")
-    shared_cluster, shared_function = shared_resources
+    current_username = rh.configs.get("username")
+    current_token = rh.configs.get("token")
 
-    with test_account():
+    with test_account:
         # Delete user access to the function
         resource_uri = rns_client.resource_uri(shared_function.rns_address)
 
@@ -156,17 +152,17 @@ def test_running_func_with_cluster_read_access():
 
 
 @pytest.mark.clustertest
-def test_running_func_with_cluster_write_access(shared_resources):
+def test_running_func_with_cluster_write_access(
+    shared_cluster, shared_function, test_account
+):
     """Check that a user with write access to a cluster can call a function on that cluster, even without having
     explicit access to the function."""
-    current_username = configs.get("username")
-    current_token = configs.get("token")
-
-    shared_cluster, shared_function = shared_resources
+    current_username = rh.configs.get("username")
+    current_token = rh.configs.get("token")
 
     cluster_uri = rns_client.resource_uri(shared_cluster.rns_address)
 
-    with test_account():
+    with test_account:
         # Give user write access to cluster from test account
         resp = requests.put(
             f"{rns_client.api_server_url}/resource/{cluster_uri}/users/access",
@@ -204,15 +200,15 @@ def test_running_func_with_cluster_write_access(shared_resources):
 
 
 @pytest.mark.clustertest
-def test_running_func_with_no_cluster_access(shared_resources):
+def test_running_func_with_no_cluster_access(
+    shared_cluster, shared_function, test_account
+):
     """Check that a user with no access to the cluster can still call a function on that cluster if they were
     given explicit access to the function."""
-    current_username = configs.get("username")
-    current_token = configs.get("token")
+    current_username = rh.configs.get("username")
+    current_token = rh.configs.get("token")
 
-    shared_cluster, shared_function = shared_resources
-
-    with test_account():
+    with test_account:
         # Delete user access to cluster using the test account
         cluster_uri = rns_client.resource_uri(shared_cluster.rns_address)
         resp = requests.delete(
